pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hello-greeter"
version = "1.0.0"
description = "A simple hello-world SDK to demonstrate pip-installable packages"
requires-python = ">=3.10"
dependencies = ["python-dotenv>=1.0"]

[project.scripts]
hello-greet = "hello_greeter.cli:main"

[tool.setuptools.packages.find]
include = ["hello_greeter*"]
